🔨 Added `restart_script.sh`: Does a complete reinstall & run of the kwin script.
There's also a wrapper script that restarts kwin
afterwards: `update_and_restart_kwin.sh`.
Caution: This starts kwin in the terminal the script
is run in - which comes with the benefit of seeing its output.
Since kwin is hard to convince to part with its output these
days, this is quite useful. But if you close your terminal
without replacing it with another kwin instance in another
shell first, e.g. like so: `kwin_x11 --replace`, it'll also
be a mess. ;)

The corresponding npm run commands are `restart` and `restart_with_kwin`.